chore: move node project into client directory
