+++ README.md
@@ -24,3 +24,13 @@ The server should have proper tests for all functionality and tests should spin 
 ## Client
 
 A minimal client scaffold is available under `client/`. Open `client/index.html` in a browser to see the placeholder interface.
+
+## Development
+
+Run `npm install` and `npm test` from within the `client` directory:
+
+```
+cd client
+npm install
+npm test
+```
